refactored to use scanner in the admin tool
