+++ cmd/admin/repair.go
@@ -15,20 +15,14 @@
 package main
 
 import (
-	"bytes"
 	"log"
 
 	"github.com/cockroachdb/errors"
-	"github.com/dolthub/dolt/go/gen/fb/serial"
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb/durable"
 	"github.com/dolthub/dolt/go/libraries/doltcore/ref"
 	"github.com/dolthub/dolt/go/store/hash"
-	"github.com/dolthub/dolt/go/store/pool"
 	"github.com/dolthub/dolt/go/store/prolly"
-	"github.com/dolthub/dolt/go/store/prolly/message"
-	"github.com/dolthub/dolt/go/store/prolly/tree"
-	"github.com/dolthub/dolt/go/store/val"
 	"github.com/dolthub/go-mysql-server/sql"
 
 	"github.com/dolthub/doltgresql/core/integrity"
@@ -45,36 +39,31 @@ type repairSummary struct {
 	InternalChunksRewritten uint64
 }
 
-// repairer rewrites prolly tree nodes whose value_address_offsets field omits references to out-of-band
-// values. Corrupt leaf nodes are re-serialized with the current serializer using their original
-// key and value tuple bytes, which recomputes the correct offsets; ancestor nodes are then rewritten to
-// reference the repaired children. Results are cached per chunk, so identical subtrees shared between
-// branches, commits, and structurally identical tables are only repaired once.
+// repairer repairs the corruption found by the integrity scanner across a database's entire commit
+// graph. The storage-level detection and node rewriting live in the integrity package (Scanner and
+// TreeRewriter, which share behavior and caches with the reporting scan and the server's startup
+// check); the repairer orchestrates them across every commit, branch, tag, and working set, updating
+// refs to the rewritten history. Commits whose content (and ancestry) is unaffected keep their
+// original hashes.
 type repairer struct {
-	db *database
-	// treeCache maps old node chunk hashes to their repaired equivalents (identity for clean subtrees).
-	treeCache     map[integrity.CacheKey]hash.Hash
-	summary       repairSummary
-	verbose       bool
-	transformLeaf leafTransform
+	db       *database
+	rewriter *integrity.TreeRewriter
+	summary  repairSummary
+	verbose  bool
 }
 
-// leafTransform inspects a leaf node and returns a replacement message, or false if the node is
-// unchanged.
-type leafTransform func(r *repairer, pm *serial.ProllyTreeNode, kd, vd *val.TupleDesc) (serial.Message, bool, error)
-
-func newRepairer(db *database, verbose bool) *repairer {
+// newRepairer returns a repairer for the given database. The rewriter is built on |sc|, so a scan
+// phase that already ran with it (e.g. the report) shares its per-chunk results with the repair.
+func newRepairer(db *database, sc *integrity.Scanner, verbose bool) *repairer {
 	return &repairer{
-		db:            db,
-		treeCache:     make(map[integrity.CacheKey]hash.Hash),
-		verbose:       verbose,
-		transformLeaf: repairLeafTransform,
+		db:       db,
+		rewriter: integrity.NewTreeRewriter(sc, db.ns),
+		verbose:  verbose,
 	}
 }
 
 // repairDatabase rewrites every commit reachable from every branch and tag of the database, repairing
 // all corrupted table data, then updates branch, tag, and working set refs to the rewritten commits.
-// Commits whose content (and ancestry) is unaffected keep their original hashes.
 func (r *repairer) repairDatabase(sctx *sql.Context) (*repairSummary, error) {
 	ddb := r.db.ddb
 	branches, err := ddb.GetBranches(sctx)
@@ -170,6 +159,8 @@ func (r *repairer) repairDatabase(sctx *sql.Context) (*repairSummary, error) {
 		}
 	}
 
+	r.summary.LeafChunksRewritten = r.rewriter.LeafChunksRewritten
+	r.summary.InternalChunksRewritten = r.rewriter.InternalChunksRewritten
 	return &r.summary, nil
 }
 
@@ -342,7 +333,7 @@ func (r *repairer) repairRootValue(sctx *sql.Context, root doltdb.RootValue) (do
 			return nil, false, err
 		}
 		oldAddr := m.Node().HashOf()
-		newAddr, err := r.repairTree(sctx, oldAddr, ti.KeyDesc, ti.ValDesc)
+		newAddr, err := r.rewriter.RewriteTree(sctx, oldAddr, ti.KeyDesc, ti.ValDesc)
 		if err != nil {
 			return nil, false, errors.Wrapf(err, "failed to repair table %s", ti.Name.String())
 		}
@@ -366,210 +357,6 @@ func (r *repairer) repairRootValue(sctx *sql.Context, root doltdb.RootValue) (do
 		changed = true
 	}
 	return root, changed, nil
-}
-
-// repairTree repairs the subtree rooted at |addr|, returning the (possibly identical) new root address.
-func (r *repairer) repairTree(sctx *sql.Context, addr hash.Hash, kd, vd *val.TupleDesc) (hash.Hash, error) {
-	key := integrity.CacheKey{Addr: addr, Desc: integrity.DescFingerprint(kd, vd)}
-	if newAddr, ok := r.treeCache[key]; ok {
-		return newAddr, nil
-	}
-
-	msg, err := integrity.GetTreeNodeMessage(sctx, r.db.cs, addr)
-	if err != nil {
-		return hash.Hash{}, err
-	}
-	var pm serial.ProllyTreeNode
-	err = serial.InitProllyTreeNodeRoot(&pm, msg, serial.MessagePrefixSz)
-	if err != nil {
-		return hash.Hash{}, err
-	}
-
-	newAddr := addr
-	if pm.TreeLevel() == 0 {
-		newMsg, changed, err := r.transformLeaf(r, &pm, kd, vd)
-		if err != nil {
-			return hash.Hash{}, errors.Wrapf(err, "failed to rewrite leaf node %s", addr.String())
-		}
-		if changed {
-			newAddr, err = r.writeNode(sctx, newMsg)
-			if err != nil {
-				return hash.Hash{}, errors.Wrapf(err, "failed to write rewritten leaf node %s", addr.String())
-			}
-			r.summary.LeafChunksRewritten++
-		}
-	} else {
-		children := integrity.ChildAddresses(&pm)
-		newChildren := make([]hash.Hash, len(children))
-		childrenChanged := false
-		for i, child := range children {
-			newChild, err := r.repairTree(sctx, child, kd, vd)
-			if err != nil {
-				return hash.Hash{}, err
-			}
-			newChildren[i] = newChild
-			childrenChanged = childrenChanged || newChild != child
-		}
-		if childrenChanged {
-			newAddr, err = r.rewriteInternal(sctx, msg, &pm, newChildren, kd, vd)
-			if err != nil {
-				return hash.Hash{}, errors.Wrapf(err, "failed to rewrite internal node %s", addr.String())
-			}
-			r.summary.InternalChunksRewritten++
-		}
-	}
-
-	r.treeCache[key] = newAddr
-	return newAddr, nil
-}
-
-// repairLeafTransform re-serializes a corrupt leaf node with the current serializer, using the node's
-// original key and value tuple bytes. The serializer recomputes value_address_offsets from the tuple
-// contents, which repairs the omitted entries. Leaf nodes with correct offsets are left unchanged.
-func repairLeafTransform(r *repairer, pm *serial.ProllyTreeNode, kd, vd *val.TupleDesc) (serial.Message, bool, error) {
-	la, err := integrity.AnalyzeLeaf(pm, kd, vd)
-	if err != nil {
-		return nil, false, err
-	}
-	if !la.Corrupt {
-		return nil, false, nil
-	}
-
-	newMsg, err := reserializeLeaf(pm, kd, vd, r.db.ns.Pool())
-	if err != nil {
-		return nil, false, err
-	}
-
-	// The rewritten node's recomputed offsets must be complete.
-	var npm serial.ProllyTreeNode
-	err = serial.InitProllyTreeNodeRoot(&npm, newMsg, serial.MessagePrefixSz)
-	if err != nil {
-		return nil, false, err
-	}
-	newLa, err := integrity.AnalyzeLeaf(&npm, kd, vd)
-	if err != nil {
-		return nil, false, err
-	}
-	if newLa.Corrupt || newLa.Stats.UnexpectedOffsets != 0 {
-		return nil, false, errors.New("rewritten leaf node still has incorrect value_address_offsets")
-	}
-	return newMsg, true, nil
-}
-
-// reserializeLeaf re-serializes a leaf node's original key and value tuple bytes with the given
-// descriptor and verifies that the result is tuple-for-tuple identical to the original.
-func reserializeLeaf(pm *serial.ProllyTreeNode, kd, vd *val.TupleDesc, pool pool.BuffPool) (serial.Message, error) {
-	keys := extractItems(pm.KeyItemsBytes(), pm.KeyOffsetsLength(), pm.KeyOffsets)
-	values := extractItems(pm.ValueItemsBytes(), pm.ValueOffsetsLength(), pm.ValueOffsets)
-
-	serializer := message.NewProllyMapSerializer(kd, vd, pool)
-	newMsg := serializer.Serialize(keys, values, nil, 0)
-
-	// Sanity-check the rewritten node: tuples must be byte-identical to the original.
-	var npm serial.ProllyTreeNode
-	err := serial.InitProllyTreeNodeRoot(&npm, newMsg, serial.MessagePrefixSz)
-	if err != nil {
-		return nil, err
-	}
-	if !bytes.Equal(npm.KeyItemsBytes(), pm.KeyItemsBytes()) || !bytes.Equal(npm.ValueItemsBytes(), pm.ValueItemsBytes()) {
-		return nil, errors.New("rewritten leaf node has different tuple bytes than the original")
-	}
-	if npm.TreeCount() != pm.TreeCount() {
-		return nil, errors.Errorf("rewritten leaf node has tree count %d, expected %d", npm.TreeCount(), pm.TreeCount())
-	}
-	return newMsg, nil
-}
-
-// rewriteInternal re-serializes an internal node, replacing its child addresses with the repaired ones.
-// Keys, subtree counts, and level are preserved from the original node.
-func (r *repairer) rewriteInternal(sctx *sql.Context, oldMsg serial.Message, pm *serial.ProllyTreeNode, newChildren []hash.Hash, kd, vd *val.TupleDesc) (hash.Hash, error) {
-	keys := extractItems(pm.KeyItemsBytes(), pm.KeyOffsetsLength(), pm.KeyOffsets)
-	if len(keys) != len(newChildren) {
-		return hash.Hash{}, errors.Errorf("internal node has %d keys but %d children", len(keys), len(newChildren))
-	}
-
-	values := make([][]byte, len(newChildren))
-	for i := range newChildren {
-		addr := newChildren[i]
-		values[i] = addr[:]
-	}
-
-	subtrees, err := subtreeCounts(oldMsg, len(newChildren))
-	if err != nil {
-		return hash.Hash{}, err
-	}
-
-	serializer := message.NewProllyMapSerializer(kd, vd, r.db.ns.Pool())
-	newMsg := serializer.Serialize(keys, values, subtrees, int(pm.TreeLevel()))
-
-	// Sanity-check the rewritten node before writing it.
-	var npm serial.ProllyTreeNode
-	err = serial.InitProllyTreeNodeRoot(&npm, newMsg, serial.MessagePrefixSz)
-	if err != nil {
-		return hash.Hash{}, err
-	}
-	if !bytes.Equal(npm.KeyItemsBytes(), pm.KeyItemsBytes()) {
-		return hash.Hash{}, errors.New("rewritten internal node has different key bytes than the original")
-	}
-	if npm.TreeLevel() != pm.TreeLevel() || npm.TreeCount() != pm.TreeCount() {
-		return hash.Hash{}, errors.New("rewritten internal node has different level or tree count than the original")
-	}
-	newAddrs := integrity.ChildAddresses(&npm)
-	if len(newAddrs) != len(newChildren) {
-		return hash.Hash{}, errors.New("rewritten internal node has wrong child count")
-	}
-	for i := range newAddrs {
-		if newAddrs[i] != newChildren[i] {
-			return hash.Hash{}, errors.New("rewritten internal node has wrong child addresses")
-		}
-	}
-
-	return r.writeNode(sctx, newMsg)
-}
-
-// writeNode writes a serialized tree node message to the node store and returns its address.
-func (r *repairer) writeNode(sctx *sql.Context, msg serial.Message) (hash.Hash, error) {
-	node, fileID, err := tree.NodeFromBytes(msg)
-	if err != nil {
-		return hash.Hash{}, err
-	}
-	if fileID != serial.ProllyTreeNodeFileID {
-		return hash.Hash{}, errors.Errorf("rewritten node has unexpected file ID %s", fileID)
-	}
-	return r.db.ns.Write(sctx, node)
-}
-
-// extractItems splits a flatbuffer item buffer into individual items using its offsets vector.
-// |offsetsLen| is the length of the offsets vector, which contains one more entry than there are items.
-func extractItems(buf []byte, offsetsLen int, offsetAt func(int) uint16) [][]byte {
-	if offsetsLen <= 1 {
-		return nil
-	}
-	items := make([][]byte, offsetsLen-1)
-	for i := 0; i < offsetsLen-1; i++ {
-		items[i] = buf[offsetAt(i):offsetAt(i+1)]
-	}
-	return items
-}
-
-// subtreeCounts decodes the subtree count array of an internal node message.
-func subtreeCounts(msg serial.Message, count int) ([]uint64, error) {
-	node, fileID, err := tree.NodeFromBytes(msg)
-	if err != nil {
-		return nil, err
-	}
-	if fileID != serial.ProllyTreeNodeFileID {
-		return nil, errors.Errorf("unexpected file ID %s", fileID)
-	}
-	node, err = node.LoadSubtrees()
-	if err != nil {
-		return nil, err
-	}
-	counts := make([]uint64, count)
-	for i := 0; i < count; i++ {
-		counts[i] = node.GetSubtreeCount(i)
-	}
-	return counts, nil
 }
 
 // commitsDiffer returns whether two commits have different hashes.
